ControlGallery Android Activity cleanup (#1011)
* Split up the various forms of Activity1 so the code is manageable

* Update access to Insights key for FormsApplicationActivity

* Remove conditional compilation symbol
