+++ Xamarin.Forms.ControlGallery.Android/DisposeLabelRenderer.cs
@@ -1,0 +1,17 @@
+using Xamarin.Forms.Controls;
+using Xamarin.Forms.Platform.Android;
+
+namespace Xamarin.Forms.ControlGallery.Android
+{
+	public class DisposeLabelRenderer : LabelRenderer
+	{
+		protected override void Dispose (bool disposing)
+		{
+
+			if (disposing) {
+				((DisposeLabel) Element).SendRendererDisposed ();
+			}
+			base.Dispose (disposing);
+		}
+	}
+}
