+++ Xamarin.Forms.ControlGallery.Android/DisposePageRenderer.cs
@@ -1,0 +1,17 @@
+using Xamarin.Forms.Controls;
+using Xamarin.Forms.Platform.Android;
+
+namespace Xamarin.Forms.ControlGallery.Android
+{
+	public class DisposePageRenderer : PageRenderer
+	{
+		protected override void Dispose (bool disposing)
+		{
+			if (disposing) {
+				((DisposePage) Element).SendRendererDisposed ();
+			}
+			base.Dispose (disposing);
+
+		}
+	}
+}
