+++ Xamarin.Forms.ControlGallery.Android/StringProvider.cs
@@ -1,0 +1,15 @@
+using Xamarin.Forms.Controls;
+
+namespace Xamarin.Forms.ControlGallery.Android
+{
+	public class StringProvider : IStringProvider
+	{
+		public string CoreGalleryTitle
+		{
+			get
+			{
+				return "Android CoreGallery";
+			}
+		}
+	}
+}
