+++ Xamarin.Forms.ControlGallery.Android/Xamarin.Forms.ControlGallery.Android.csproj
@@ -175,11 +175,19 @@
     <Compile Include="Activity1.cs" />
     <Compile Include="BorderEffect.cs" />
     <Compile Include="BrokenNativeControl.cs" />
+    <Compile Include="CacheService.cs" />
+    <Compile Include="FormsApplicationActivity.cs" />
+    <Compile Include="DisposeLabelRenderer.cs" />
+    <Compile Include="DisposePageRenderer.cs" />
+    <Compile Include="FormsAppCompatActivity.cs" />
+    <Compile Include="PreApplicationClassActivity.cs" />
     <Compile Include="Properties\MapsKey.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Resources\Resource.Designer.cs" />
     <Compile Include="CustomRenderers.cs" />
     <Compile Include="ColorPicker.cs" />
+    <Compile Include="StringProvider.cs" />
+    <Compile Include="TestCloudService.cs" />
     <Compile Include="_38989CustomRenderer.cs" />
     <Compile Include="BrokenImageSourceHandler.cs" />
     <Compile Include="_57114CustomRenderer.cs" />
